+++ app.js
@@ -828,6 +828,7 @@ import { renderers as RV1472 } from './views/group-v1472.js';
 import { renderers as RV1473 } from './views/group-v1473.js';
 import { renderers as RV1475 } from './views/group-v1475.js';
 import { renderers as RV1476 } from './views/group-v1476.js';
+import { renderers as RV1479 } from './views/group-v1479.js';
 import { renderers as RV1396 } from './views/group-v1396.js';
 import { renderers as RV164 } from './views/group-v164.js';
 import { renderers as RV165 } from './views/group-v165.js';
@@ -916,7 +917,7 @@ const RENDERERS = { ...RA, ...RB, ...RC, ...RE, ...RF, ...RG, ...RH, ...RI, ...R
   ...RV902,
   ...RV903,
   ...RV905,
-  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV1475, ...RV1476, ...RV63, ...RPALINT };
+  ...RV906, ...RV907, ...RV908, ...RV909, ...RV910, ...RV911, ...RV912, ...RV916, ...RV917, ...RV918, ...RV919, ...RV920, ...RV921, ...RV922, ...RV923, ...RV924, ...RV925, ...RV927, ...RV928, ...RV932, ...RV958, ...RV960, ...RV1061, ...RV1062, ...RV1240, ...RV1241, ...RV1242, ...RV1243, ...RV1400, ...RV1389, ...RV1393, ...RV1397, ...RV1395, ...RV1390, ...RV1391, ...RV1392, ...RV1394, ...RV1399, ...RV1398, ...RV1401, ...RV1396, ...RV1412, ...RV1413, ...RV1414, ...RV1416, ...RV1417, ...RV1418, ...RV1419, ...RV1420, ...RV1421, ...RV1422, ...RV1423, ...RV1424, ...RV1425, ...RV1426, ...RV1427, ...RV1428, ...RV1429, ...RV1430, ...RV1431, ...RV1433, ...RV1434, ...RV1435, ...RV1436, ...RV1438, ...RV1439, ...RV1440, ...RV1441, ...RV1442, ...RV1443, ...RV1444, ...RV1445, ...RV1446, ...RV1447, ...RV1448, ...RV1449, ...RV1450, ...RV1451, ...RV1452, ...RV1453, ...RV1454, ...RV1455, ...RV1468, ...RV1469, ...RV1470, ...RV1471, ...RV1472, ...RV1473, ...RV1475, ...RV1476, ...RV1479, ...RV63, ...RPALINT };
 
 // ----- Utility registry ----------------------------------------------------
 // Source of truth for routes, names, group, audiences, and clinical flag.
@@ -1846,6 +1847,7 @@ const UTILITIES = [
   { id: 'samilson-prieto',          name: 'Samilson-Prieto Grade (Shoulder Dislocation Arthropathy)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'rockwood-ac',              name: 'Rockwood Classification (Acromioclavicular Joint Injury)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'neer-distal-clavicle', name: 'Modified Neer Classification (Distal Clavicle Fractures)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
+  { id: 'robinson-clavicle', name: 'Robinson Classification (Clavicle Fractures)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'bigliani-acromion',        name: 'Bigliani Classification (Acromion Morphology)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'fernandez-radius',         name: 'Fernandez Classification (Distal Radius Fracture)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },
   { id: 'ruedi-allgower-pilon',     name: 'Ruedi-Allgower Classification (Tibial Pilon Fracture)', group: 'G', audiences: ['clinicians', 'educators'], clinical: true },